+++ backend/routes/loginAuth.ts
@@ -1,11 +1,6 @@
 import express from "express";
 const loginAuthRouter = express.Router()
-import { Strategy as LocalStrategy } from 'passport-local';
 import passport from 'passport'
-import prisma from '../utils/prismaClient'
-import crypto from 'crypto';
-
-const db = prisma;
 
 loginAuthRouter.get('/login/password',
     function (req, res, next) {
@@ -19,28 +14,5 @@ loginAuthRouter.post('/login/password',
         res.status(200).send("we logged in")
     });
 
-passport.use(new LocalStrategy(async (email: string, password: string, done: Function) => {
-    try {
-        const user = await db.user.findUnique({
-            where: {
-                email
-            },
-        });
-
-        if (!user || !user.salt) {
-            return done(null, false, { message: 'Incorrect email or password.' });
-        }
-
-        const hashedPassword = crypto.pbkdf2Sync(password, user.salt, 310000, 32, 'sha256').toString('hex');
-        if (user.password !== hashedPassword) {
-            return done(null, false, { message: 'Incorrect email or password.' });
-        }
-
-        return done(null, user);
-
-    } catch (error) {
-        done(error);
-    }
-}));
 
 export default loginAuthRouter
